Extracted IPP to HAL for calcHist function
